svc_xprt_lookup - Add extra ref on create
An xprt has a ref for the hash table (that's released by SVC_DESTROY());
but when it's first created, only 1 ref was taken, so there wasn't a ref
for the caller.

Add an extra ref for the caller when the xprt is first created.

Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
(cherry picked from commit ca74cde10ef02a322b8944a6c8639b1318fa34dc)