Destroy XPRTs for local clients
When destroying a client, the XPRT is not also destroyed, but just
unref'd.  This is fine, if multiple clients share a XPRT, but local
clients all have unique FDs, and therefore unique XPRTs.  This means
that their XPRTs are never destroyed.  Ganesha uses these a lot, leading
to many leaked XPRTs and therefore FDs.

Add a flag on a client that indicates it's local, and use that flag to
desrtoy the XPRT when the client is destroyed.

Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
(cherry picked from commit b86186c0f7654580e615d2386d6eead87e9fa345)