getclnthandle(): Fix assert(client == NULL) and destroy client handle.
In getclnthandle() function  'clnt_tli_ncreate()' is called. Currently
even if 'clnt_tli_ncreate()' doesn't complete successfully it returns a
valid pointer for 'client' and the status is saved in
'client->cl_error.re_status'. But with this later on in getclnthandle()
ganesha may abort for condition assert(client == NULL).
To fix this we now check if 'client' is not NULL then
assert(CLNT_FAILURE(client)) and then destroy the client.

Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>
(cherry picked from commit 95155fa201421315e629788b719580357053037f)