svc_dg.c: cleanup pktinfo handling
Prior to this change it was seen that the MSG_CTRUNC flag was set in
the data returned by recvmsg(), an indication that that the controlmsg
(i.e. pktinfo) was lost due to insufficient space. Increasing the size
of struct svc_dg_xprt by making sc_cmsg an array of pktinfo fixes this.

It was also seen that when using an IPv4 socket that two pktinfos are
recv'd: an in_pktinfo and an in6_pktinfo. (One comment I saw somewhere
said that it is related to ipv6-tunneled-over-ipv4.) In this change the
in6_pktinfo is ignored/discarded and only the in_pktinfo is saved for
later use in the reply.

Presumably an IPv6 socket receives just a single in6_pktinfo, which is
saved.

Lastly, despite a comment that seemed to imply that pktinfo is/was set
in the reply, there was no logic to actually set it; with the result
that the kernel had no hints about which source address to set in the
packet and defaulting to sending it with the lowest numeric IP address
on the NIC. This may or may not have been the correct address, i.e.
the address that the request was actually sent to. Discussions with
the knfs team confirmed that NFS clients would consider this an error.

And it was seen that the client's network stack could reject such
packets in certain circumstances, and the NFS client in failing to
receive such a reply would resend the requests (with those replies
rejected as well) before eventually failing over to sending TCP
requests as a (last ditch) attempt to resolve the mount.

(cherry picked from commit 96018f8fcd4194cb70e9b4636f918e807d3b5e84)