XPRT - Unregister xprt before dropping sentinal ref
XPRTs are generally stored in some form of lookup table.  When we're
destroying the XPRT, we need to remove it from it's lookup table before
we drop the sentinal ref.  Otherwise, another thread can find it, and
increment it's ref from 0 -> 1.  This is very hard to handle in an
atomic fashion.

Add an unlink() API method, and implement it for all backends.  This
allows us to remove the XPRT from it's lookup table before we drop it's
sentinal ref.

Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
(cherry picked from commit eed2cc7999d3f5ad1c1e512706622fde1c05713e)